Report TLS 1.3 as supporting secure renegotiation.
TLS 1.3 doesn't support renegotiation in the first place, but so callers
don't report TLS 1.3 servers as missing it, always report it as
(vacuously) protected against this bug.

(Chrome on Android still has remnants of some UI that reports servers
missing this. We should get rid of that, but I think this is reasonable
regardless. I expect we've lost the tokbind NAND renego game, but it'll
instead require this extension.)

Change-Id: Ibfec03102b2aec7eaa773c331d6844292e7bb685