Update patch set 3
Patch Set 3: Code-Review-1

Even with the Android stuff fixed, there's a lot of code that stack-allocates BIGNUMs, not to mention the efficiency gains from being able to do so or embed them in structs. I doubt we'll be able to do this anytime soon, if ever.

If we wanted to get rid of direct access to the struct, one option is to rename all the members, which ensures that existing code that reaching into it breaks. But I don't think we're ready to do that yet either.

Patch-set: 3
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=-1