Update patch set 2
Patch Set 2:

To my knowledge, every device with an ARMv8 chip ships with a working getauxval, so there's no need to poke at /proc/self/auxv. Note that cpu-aarch64-linux.c, used in the AArch64 mode for each of these devices, doesn't do any of this nonsense and even links to getauxval in the usual way.

However, some of those devices do not put AT_HWCAP2 in the auxv at all, needed for detecting ARMv8 features in AArch32 mode. Apparently this is a separate kernel config, and they forgot to toggle it?! This is orthogonal from how to get at the auxv, so AT_HWCAP2 will be missing from both getauxval and /proc/self/auxv. However, because /proc/cpuinfo is not based on auxv and is just what features the kernel believes exists, that'll work independent of that.

So I think the only way this would be needed would be if there are devices where AArch64 builds of BoringSSL can't even run at all right now (no weak symbol) and no one noticed or if somehow only their 64-bit libc has a getauxval and not the 32-bit one.

(Come to think of it, I should expose more booleans for these so we can get metrics in Chrome and maybe retire this someday. At least one of the affected devices is a Nexus device, so there's some vague semblance of a hope that it'll get upgraded away. Though BoringSSL is used in Android apps beyond Chrome, so I don't have a sense of what threshold would be considered acceptable.)

Patch-set: 2