Update patch set 2
Patch Set 2:

Oh, and the reason not to always use /proc/cpuinfo is because of the comment about optional features here.
https://android.googlesource.com/platform/ndk/+/882ac8f3392858991a0e1af33b4b7387ec856bd2

So the assumption here is that, for each bug that causes us to fallback to /proc/cpuinfo, devices shipping with the next architecture revision will have the fix. (Though so far they keep managing to mess it up!)

Well, that and parsing strings is silly. :-)

Patch-set: 2