Update patch set 1
Patch Set 1:

(Confirmed good CLA.)

Thanks for this. As you've probably noticed, we're in the process of doing a bunch of reworking of certificate handling in order to be able to drop crypto/x509 and to save memory in clients.

Adding cleaner APIs would be one of the later stages of this process but the thing I currently had in mind was something like:

int SSL_CTX_set_cert_chain(CRYPTO_BUFFER **bufs, size_t num_bufs);

I.e. set the whole lot at once with a function that takes references to CRYPTO_BUFFERS. (So that, if you have many SSL_CTXs, they can all share intermediate certificates.)

Would that work for you? (It might be a few weeks until it's ready however.)

Patch-set: 1