Permit AEAD ciphers for encrypting session tickets.
This change allows session tickets to be encrypted with AEAD ciphers
instead of the current cipher + HMAC scheme. The ticket callback
returns the special error value |TLSEXT_TICKET_CB_WANT_AEAD| to
request the use of an AEAD cipher. When this value is returned the
ticket callback is called a second time with a |EVP_AEAD_CTX|* passed
in as |ctx|, and a magic value passed in as |hmac_ctx|.

This isn't the cleanest API, but it is not dissimilar to the
|SSL_magic_pending_session_ptr| API.

This is here for comments and review, but is probably not ready to be
merged.

Change-Id: Ib564fa98db9b888e4c5e6fc8c194ca06f1de8eed