Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> > (1 comment)
> 
> Thanks for thinking about me :) Actually, I have some good news. We are probably going to be able to remove this whole custom session timeouts mess from our code so not only we won't need this, but I can also revert my previous change if that's something that would make you happy (sorry for this whole mess).
>
> On the other hand we'll still need similar functions for e.g. setting SCTs and enabling TLS 1.3 0RTT, so I guess it kind of balances out.

Ah. Well, if you don't need it either, we may as well remove it. :-) Fewer public APIs means fewer knobs to test.

Although if one's deployment has shorter ticket key lifetimes than the hard-coded value, that's a good reason to make the TLS 1.3 thing configurable, to avoid known-useless 0-RTT offers. *shrug* It's not a particularly offensive configuration knob or anything.

(No worries. It got us to finally come up with a per-SSL config plan, so that was good. Still need to actually implement that plan, but I think we were more worried about not having a way out than being imminently in need of it.)

Patch-set: 6