Append newline to line param of SSL_CTX_get_keylog_callback.
Libraries might want to write the keylog using one write() syscall as a
synchronization method. By providing the newline in the callback, the
callbacks won't have to make a copy of the buffer (or resort to more
complicated constructions using writev).

POSIX guarantees an atomic write for the full buffer (smaller than
PIPE_BUF >= 512) for pipes (like /dev/stderr) and FIFOs. For Linux, the
main concern is interruption of write by signals (see write(2)), but
signal(7) claims that I/O operations on (local) disks are not
interrupted by signals so indeed a single write should be OK.

Risk for existing consumers: basically none. Existing consumers may add
an additional newline, but that is ignored by Wireshark.

Change-Id: Ic0ebc00655df1007c0d73e91443004b2e29d32d0