Append newline to line param of SSL_CTX_get_keylog_callback.
Libraries might want to write the keylog using one write() syscall as a
synchronization method. By providing the newline in the callback, the
callbacks won't have to make a copy of the buffer.

Risk for existing consumers: basically none. Existing consumers may add
an additional newline, but that is ignored by Wireshark.

Change-Id: Ic0ebc00655df1007c0d73e91443004b2e29d32d0