Update patch set 1
Patch Set 1:

Not quite sure if this was the outcome of the discussion, so do with this whatever you want.

I'm also not sure if initial_ctx is the correct ctx to use here, but the code I originally modified used that.

Patch-set: 1
Reviewer: Gerrit User 7060 <7060@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0