Update patch set 1
Patch Set 1:

(CLA confirmed.)

It probably wouldn't take much to update users of this, but doesn't writev solve this problem?

(Perhaps writing these to a text file is such a common case that we should optimise for it anyway. Just not immediately sure if changing the semantics of this is best.)

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0