Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (CLA confirmed.)
> 
> It probably wouldn't take much to update users of this, but doesn't writev solve this problem?
> 
> (Perhaps writing these to a text file is such a common case that we should optimise for it anyway. Just not immediately sure if changing the semantics of this is best.)

At the very least, it would need to come with a BORINGSSL_API_VERSION bump so we can coordinate this with consumers properly. It's a fairly trivial update to downstreams, but does seem like a waste of time.

(Is a single call to write() actually guaranteed to be atomic in POSIX? If not, the motivation here is a little dubious.)

Patch-set: 1
Label: Code-Review=0