Update patch set 2
Patch Set 2: Code-Review-1

> Patch Set 2:
> 
> (Rebased and updated commit message to (hopefully) address the concerns about guarantees of write().)

This still seems iffy.

Atomicity guarantees for a pipe aren't all that important here; applications will typically log it to a file.

Does POSIX promise that two threads calling write() to a file will not interleave? That's not quite the same as promising that write() won't be interrupted by a signal. Suppose things were aligned on blocks funny such that I only got half a line into some buffer and the disk were busy. Does the kernel promise that it won't write something else when it wakes up again? I've never heard of an application relying on this. Usually you add a lock or funnel things into a single thread.

> The change in the interface is quite minor and normally the additional newline should not cause issues. Is it worth bumping the API number for this?

Yes. Because the before and after stages are incompatible with each other and you are changing a function that is in use, we would need some preprocessor hook to coordinate the change in consumers. We would then hold back the change until all consumers have been updated.

Though, honestly, I think this API is both worse for our purposes (WebRTC in a separate process, etc.) and the issue is sufficiently trivial that this is basically a waste of time. We can change API, but it's not zero cost.

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=-1