Remove unused RSAZ AVX2 code.
Since |$avx| is set to zero, |rsaz_avx2_eligible| will always return 0,
so this is dead code.

In private discussion I was told that Haswell is the only platform for
which the the RSAZ AVX2 code is faster than the integer code. On
earlier CPUs, AVX2 isn't available. On later CPUs, ADX is available,
and the ADX-based code is faster, according to the performance notes in
the OpenSSL version of rsaz-avx2.pl. It's not worthwhile to carry this
code just for Haswell.

Change-Id: If8d8a10f2420b5d3fc22f83aec9365834810bb90