Update patch set 3
Patch Set 3:

Obviously there might be reasons for keeping this code around that I'm not aware of, that might make this code more interesting than I suspect it is. But if not, it's good to get rid of it before any effort is spent on enabling it.

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>