Don't use windowed method for variable-time exponentiation.
Assuming that the performance of variable-time Montgomery
exponentiation is only relevant for RSA, and assuming that the
(public) exponent is always small, simplify |BN_mod_exp_mont| to always
use the non-windowed square-and-multiply algorithm. Also, reduce the
maximum amount of memory used by this function.

Previously, the windowed algorithm would be used (only) for exponents of
at least 24 bits, as explained in the comment above
|BN_window_bits_for_exponent_size|. However, in the case of RSA, the
public exponent is almost always 65537 or 3, which don't meet that bar.
65537 and 3 have Hamming weight 2, and the non-windowed
square-and-multiply method is optimal for an exponent with Hamming
weight 2.

Also note that |check_modulus_and_exponent_sizes| limits the public
exponent to at most 33 bits; this limit helps minimize the worst-case
negative performance impact of this change.

Change-Id: I545088ec0138f4f34eb6da3abca4858fe7757053