Push password encoding back into pkcs12_key_gen.
With PKCS8_encrypt_pbe and PKCS8_decrypt_pbe gone in
3e8b782c0cc0d9621f622cf80ab1a9bcf442fa17, we can restore the old
arrangement where the password encoding was handled in pkcs12_key_gen.
This simplifies the interface for the follow-up crypto/asn1 split.

Note this change is *not* a no-op for PKCS#12 files which use PBES2.
Before, we would perform the PKCS#12 password encoding for all parts of
PKCS#12 processing. The new behavior is we only perform it for the parts
that go through the PKCS#12 KDF. For such a file, it would only be the
MAC.

I believe the specification supports our new behavior. Although RFC 7292
B.1 says something which implies that the transformation is about
converting passwords to byte strings and would thus be universal,
appendix B itself is prefaced with:

   Note that this method for password privacy mode is not recommended
   and is deprecated for new usage.  The procedures and algorithms
   defined in PKCS #5 v2.1 [13] [22] should be used instead.
   Specifically, PBES2 should be used as encryption scheme, with PBKDF2
   as the key derivation function.

"This method" refers to the key derivation and not the password
formatting, but it does give support to the theory that password
formatting is tied to PKCS#12 key derivation.

(Of course, if one believes PKCS#12's assertion that their inane
encoding (NUL-terminated UTF-16!) is because PKCS#5 failed to talk about
passwords as Unicode strings, one would think that PBES2 (also in
PKCS#5) would have the same issue and thus need PKCS#12 to valiantly
save the day with an encoding...)

Anyway, the reality is PKCS#12 is a disaster and must be defined via
checking other implementations. On that front, OpenSSL aligns with our
new behavior. (Which makes sense since our "new" behavior is really
reverting a change some time ago where we accidentally diverged.) Today,
we are unable to process PBES2 PKCS#12 files that OpenSSL produces and,
with this change, we are.

However, NSS and macOS does not align with this. They cannot process
such files, even though they appear to have code for it. I'm not sure
what variant, if any, they *do* process as I haven't been able to
produce one which satisfies them. Absent a coherent story here, let's
just try to match OpenSSL as we expose OpenSSL-ish APIs.

TODO - Try a bit harder to figure out what's going on with at least one
other implementation and then give up... maybe gdb into NSS a bit?

BUG=54

Change-Id: I9a1bb4d5e168e6e76b82241e4634b1103e620b9b